Fixed Sqlite floating point values reading

+++ dotmimsync/src/main/java/com/mimetis/dotmim/sync/sqlite/CursorHelper.kt
@@ -2,18 +2,14 @@ package com.mimetis.dotmim.sync.sqlite
 
 import android.database.Cursor
 import android.database.Cursor.*
-import androidx.core.database.getBlobOrNull
-import androidx.core.database.getFloatOrNull
-import androidx.core.database.getIntOrNull
-import androidx.core.database.getStringOrNull
-import com.mimetis.dotmim.sync.setup.DbType
+import androidx.core.database.*
 
 object CursorHelper {
     fun Cursor.getValue(columnIndex: Int): Any? =
             when (this.getType(columnIndex)) {
                 FIELD_TYPE_NULL -> null
                 FIELD_TYPE_INTEGER -> this.getIntOrNull(columnIndex)
-                FIELD_TYPE_FLOAT -> this.getFloatOrNull(columnIndex)
+                FIELD_TYPE_FLOAT -> this.getDoubleOrNull(columnIndex)
                 FIELD_TYPE_STRING -> this.getStringOrNull(columnIndex)
                 FIELD_TYPE_BLOB -> this.getBlobOrNull(columnIndex)
                 else -> null
